requirements.txt
numpy
sklearn
antspyx
torch
kornia
transforms3d